# api-docs.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "Katron AI Gift Card API Documentation",
                 "version":  "1.0"
             },
    "servers":  [
                    {
                        "url":  "https://api.ktngiftcard.katronai.com/katron-gift-card",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "BearerAuth":  [

                                        ]
                     }
                 ],
    "paths":  {
                  "/api/user/uploadGenericFile":  {
                                                      "post":  {
                                                                   "tags":  [
                                                                                "user-controller"
                                                                            ],
                                                                   "operationId":  "uploadGenericFile",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "folderName",
                                                                                          "in":  "query",
                                                                                          "required":  true,
                                                                                          "schema":  {
                                                                                                         "type":  "string"
                                                                                                     }
                                                                                      },
                                                                                      {
                                                                                          "name":  "fileName",
                                                                                          "in":  "query",
                                                                                          "required":  true,
                                                                                          "schema":  {
                                                                                                         "type":  "string"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "requestBody":  {
                                                                                       "content":  {
                                                                                                       "multipart/form-data":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "required":  [
                                                                                                                                                                   "file"
                                                                                                                                                               ],
                                                                                                                                                  "type":  "object",
                                                                                                                                                  "properties":  {
                                                                                                                                                                     "file":  {
                                                                                                                                                                                  "type":  "string",
                                                                                                                                                                                  "format":  "binary"
                                                                                                                                                                              }
                                                                                                                                                                 }
                                                                                                                                              }
                                                                                                                               }
                                                                                                   },
                                                                                       "required":  true
                                                                                   },
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "OK",
                                                                                                 "content":  {
                                                                                                                 "*/*":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/Response"
                                                                                                                                        }
                                                                                                                         }
                                                                                                             }
                                                                                             }
                                                                                 }
                                                               }
                                                  },
                  "/api/user/updateUserPhoto":  {
                                                    "post":  {
                                                                 "tags":  [
                                                                              "user-controller"
                                                                          ],
                                                                 "operationId":  "updateUserPhoto",
                                                                 "parameters":  [
                                                                                    {
                                                                                        "name":  "photo",
                                                                                        "in":  "query",
                                                                                        "required":  true,
                                                                                        "schema":  {
                                                                                                       "type":  "string"
                                                                                                   }
                                                                                    }
                                                                                ],
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  "OK",
                                                                                               "content":  {
                                                                                                               "*/*":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/Response"
                                                                                                                                      }
                                                                                                                       }
                                                                                                           }
                                                                                           }
                                                                               }
                                                             }
                                                },
...
                  "/api/user/updateUserInformation":  {
                                                          "post":  {
                                                                       "tags":  [
                                                                                    "user-controller"
                                                                                ],
                                                                       "operationId":  "updateUserInformation",
                                                                       "requestBody":  {
                                                                                           "content":  {
                                                                                                           "application/json":  {
                                                                                                                                    "schema":  {
                                                                                                                                                   "$ref":  "#/components/schemas/UpdateUserModel"
                                                                                                                                               }
                                                                                                                                }
                                                                                                       },
                                                                                           "required":  true
                                                                                       },
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "OK",
                                                                                                     "content":  {
                                                                                                                     "*/*":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/Response"
                                                                                                                                            }
                                                                                                                             }
                                                                                                                 }
                                                                                                 }
                                                                                     }
                                                                   }
                                                      },
...
              },
    "components":  {
                       "schemas":  {
                                       "Response":  {
                                                        "type":  "object",
                                                        "properties":  {
                                                                           "status":  {
                                                                                          "type":  "integer",
                                                                                          "format":  "int32"
                                                                                      },
                                                                           "message":  {
                                                                                           "type":  "string"
                                                                                       },
                                                                           "data":  {
                                                                                        "type":  "object"
                                                                                    }
                                                                       }
                                                    },
                                       "UpdateUserModel":  {
                                                               "required":  [
                                                                                "firstname",
                                                                                "lastname",
                                                                                "phone"
                                                                            ],
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "id":  {
                                                                                             "type":  "integer",
                                                                                             "format":  "int64"
                                                                                         },
                                                                                  "firstname":  {
                                                                                                    "type":  "string"
                                                                                                },
                                                                                  "lastname":  {
                                                                                                   "type":  "string"
                                                                                               },
                                                                                  "phone":  {
                                                                                                "type":  "string"
                                                                                            },
                                                                                  "about":  {
                                                                                                "type":  "string"
                                                                                            }
                                                                              }
                                                           },
...
                                   },
                       "securitySchemes":  {
                                               "BearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}